app/requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
autopep8==1.7.0
bleak==0.20.2
buildozer==1.5.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
Cython==0.29.33
dbus-fast==1.85.0
distlib==0.3.6
docutils==0.18.1
Faker==14.1.0
filelock==3.12.0
firebase-admin==5.3.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
Jinja2==3.1.2
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.22.4
packaging==23.1
pep517==0.6.0
pexpect==4.8.0
Pillow==9.1.1
platformdirs==3.5.0
plyer==2.0.0
proto-plus==1.22.1
protobuf==4.21.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
Pygments==2.12.0
pyjnius==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytoml==0.1.21
requests==2.27.1
rsa==4.9
sh==1.14.2
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
toml==0.10.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.23.0
zipp==3.8.0
